Config Mgr: Provide a way to archive and restore critical non-ROS config files
NE-89 NE-43

+++ scripts/config_mgr.py
@@ -24,10 +24,31 @@ USER_SUFFIX = '.user'
 
 pending_nodes = {}
 
+# Files outside the normal NEPI-ROS cfg. scheme
+SYS_CFGS_TO_PRESERVE = {
+    'sys_env.bash' : '/opt/nepi/sys_env.bash', # Serial number, ROS launch file, external ROS MASTER etc.
+    'authorized_keys' : '/opt/nepi/config/home/nepi/ssh/authorized_keys', # NEPI Device SSH public keys
+    'hostname' : '/opt/nepi/config/etc/hostname', # NEPI Device hostname
+    'sshd_config' : '/opt/nepi/config/etc/ssh/sshd_config', # SSH Server Config
+    'chrony.conf' : '/opt/nepi/config/etc/chrony/chrony.conf.user', # NTP/Chrony Config
+    'iptables.rules' : '/opt/nepi/config/etc/iptables/s2x_iptables.rules', # Route and forwarding rules; e.g., for dual-interface devices
+    'sensor_if_static_ip' : '/opt/nepi/config/etc/network/interfaces.d/s2x_sensor_if_static_ip', # Static IP address for secondary/sensor ethernet interface
+    'ip_aliases' : '/opt/nepi/config/etc/network/interfaces.d/nepi_user_ip_aliases', # IP alias addresses for primary ethernet interface
+    'static_ip' : '/opt/nepi/config/etc/network/interfaces.d/nepi_static_ip', # Principal static IP address for primary ethernet interface
+    'fstab' : '/opt/nepi/config/etc/fstab', # Filesystem mounting rules; e.g., nepi_storage on SD vs SSD
+    'smb.conf' : '/opt/nepi/config/etc/samba/smb.conf', # Samba configuration
+    'nepi_connect_cfg' : '/opt/nepi/nepi_link/nepi-bot/devinfo/', # NEPI Connect Device NUID, SSH keys, etc.
+    'nepi_connect_bot_cfg.json' : '/opt/nepi/nepi_link/nepi-bot/cfg/bot/config.json' # NEPI Connect device (nepi-bot) config file
+}
+
 # Moving symlinks is typically faster and more robust than copying files, so to reduce the
 # chance of filesystem corruption in the event of e.g., power failure, we use a symlink-based config
 # file scheme.
 def symlink_force(target, link_name):
+    link_dirname = os.path.dirname(link_name)
+    if not os.path.exists(link_dirname):
+        rospy.logwarn("Skipping symlink for " + link_name + " because path does not exist... missing factory config?")
+        return False
     try:
         os.symlink(target, link_name)
     except OSError as e:
@@ -35,7 +56,7 @@ def symlink_force(target, link_name):
             os.remove(link_name)
             os.symlink(target, link_name)
         else:
-            rospy.logerr("Unable to create symlink %s for %s", link_name, target)
+            rospy.logerr("Unable to create symlink %s for %s: (%s)", link_name, target, str(e))
             return False
     return True
 
@@ -60,7 +81,13 @@ def get_cfg_pathname(qualified_node_name):
 
 def get_user_cfg_pathname(qualified_node_name):
     node_name = separate_node_name_in_msg(qualified_node_name)
-    user_cfg_pathname = USER_CFG_PATH + '/' + node_name + CFG_SUFFIX + USER_SUFFIX
+    user_cfg_dirname = os.path.join(USER_CFG_PATH, 'ros')
+    
+    # Ensure the path we report actually exists
+    if not os.path.isdir(user_cfg_dirname):
+        os.makedirs(user_cfg_dirname)
+
+    user_cfg_pathname = os.path.join(user_cfg_dirname, node_name + CFG_SUFFIX + USER_SUFFIX)
     return user_cfg_pathname
 
 def user_reset(req):
@@ -91,25 +118,47 @@ def store_params(msg):
 
     # Now, ensure the link points to the correct file
     cfg_pathname = get_cfg_pathname(qualified_node_name)
-    if (False == symlink_force(user_cfg_pathname, cfg_pathname)):
-        rospy.logerr("Unable to update the cfg. file link")
+    symlink_force(user_cfg_pathname, cfg_pathname) # Error logged upstream
+
+def save_non_ros_cfgs(msg):
+    target_dir = os.path.join(USER_CFG_PATH, 'sys')
+    if not os.path.exists(target_dir):
+        os.makedirs(target_dir)
+
+    for cfg in SYS_CFGS_TO_PRESERVE:
+        source = SYS_CFGS_TO_PRESERVE[cfg]
+        target = os.path.join(USER_CFG_PATH, 'sys', cfg)
+        os.system('cp -rf ' + source + ' ' + target)
 
 def restore_user_cfgs(msg):
+    # First handle the NEPI-ROS user configs.
     for root, dirs, files in os.walk(CFG_PATH):
         for name in files:
             full_name = os.path.join(root, name)
             if full_name.endswith(CFG_SUFFIX) and os.path.islink(full_name):
-                user_cfg_name = os.path.join(USER_CFG_PATH, name + USER_SUFFIX)
+                user_cfg_name = os.path.join(USER_CFG_PATH, 'ros', name + USER_SUFFIX)
                 if os.path.exists(user_cfg_name): # Restrict to those with present user configs
                     link_name = os.path.join(root, name.replace(FACTORY_SUFFIX, ''))
                     rospy.loginfo("Updating " + link_name + " to user config")
                     symlink_force(user_cfg_name, link_name)
+
+    # Now handle non-ROS user system configs.        
+    for name in SYS_CFGS_TO_PRESERVE:
+        full_name = os.path.join(USER_CFG_PATH, name)
+        if os.path.exists(full_name):
+            if os.path.isdir(full_name):
+                full_name = os.path.join(full_name,'*') # Wildcard avoids copying source folder into target folder as a subdirectory
+            target = SYS_CFGS_TO_PRESERVE[name]
+            rospy.loginfo("Updating " + target + " from user config")
+            os.system('cp -rf ' + full_name + ' ' + target)
+            os.system('chown -R nepi:nepi ' + target)
 
 def config_mgr():
     rospy.init_node('config_mgr')
 
     rospy.loginfo('Starting the config. mgr node')
 
+    rospy.Subscriber('save_config', Empty, save_non_ros_cfgs) # Global one only
     rospy.Subscriber('store_params', String, store_params)
     rospy.Subscriber('full_user_restore', Empty, restore_user_cfgs)
